--- cs251tk/lib/__init__.py
@@ -1,15 +1,0 @@
-'''Library of functions that create the toolkit'''
-
-from .args import process_args
-# from .check_for_updates import check_for_updates
-from .columnize import tabulate
-from .format_collected_data import format_collected_data
-from .helpers import chdir
-from .helpers import flatten
-from .helpers import warn
-from .gist import post_gist
-from .progress_bar import progress_bar
-from .run import run
-from .save_recordings import save_recordings
-from .single_student import single_student
-from .specs import load_specs
