+++ cs251tk/specs/__init__.py
@@ -1,0 +1,2 @@
+from .specs import load_specs
+from .specs import get_filenames
