+++ cs251tk/student/markdownify.py
@@ -1,11 +1,12 @@
 import os
 import shlex
-from glob import glob
 from collections import OrderedDict
+from glob import glob
 from os.path import exists, join as path_join
-from .find_unmerged_branches_in_cwd import find_unmerged_branches_in_cwd
-from .run import run
-from .helpers import flatten
+
+from cs251tk.common import find_unmerged_branches_in_cwd
+from cs251tk.common import flatten
+from cs251tk.common import run
 
 
 def unicode_truncate(string, length, encoding='utf-8'):
@@ -14,7 +15,7 @@ def unicode_truncate(string, length, encoding='utf-8'):
 
 
 def expand_chunk(command_chunk):
-    '''Take a chunk of a command and expand it, like a shell'''
+    """Take a chunk of a command and expand it, like a shell"""
     # TODO: Support escaped globs
     if '*' in command_chunk:
         return glob(command_chunk)
@@ -22,7 +23,7 @@ def expand_chunk(command_chunk):
 
 
 def process_chunk(command):
-    '''Takes one piece of a pipeline and formats it for run_command'''
+    """Takes one piece of a pipeline and formats it for run_command"""
     # decode('unicode_escape') de-escapes the backslash-escaped strings.
     # like, it turns the \n from "echo Hawken \n 26" into an actual newline,
     # like a shell would.
